Move main logic to internal
