--- cmd/gomplate/logger.go
@@ -1,34 +1,0 @@
-package main
-
-import (
-	"context"
-	stdlog "log"
-	"os"
-	"time"
-
-	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/log"
-	"golang.org/x/term"
-)
-
-func initLogger(ctx context.Context) context.Context {
-	// default to warn level
-	zerolog.SetGlobalLevel(zerolog.WarnLevel)
-	zerolog.DurationFieldUnit = time.Second
-
-	stdlogger := log.With().Bool("stdlog", true).Logger()
-	stdlog.SetFlags(0)
-	stdlog.SetOutput(stdlogger)
-
-	if term.IsTerminal(int(os.Stderr.Fd())) {
-		log.Logger = log.Output(zerolog.ConsoleWriter{Out: os.Stderr, TimeFormat: "15:04:05"})
-		noLevelWriter := zerolog.ConsoleWriter{
-			Out:         os.Stderr,
-			FormatLevel: func(i interface{}) string { return "" },
-		}
-		stdlogger = stdlogger.Output(noLevelWriter)
-		stdlog.SetOutput(stdlogger)
-	}
-
-	return log.Logger.WithContext(ctx)
-}
